transferre function from hooks to doctype class
