+++ one_fm/hooks.py
@@ -357,9 +357,6 @@ doc_events = {
 	"User": {
 		"after_insert":"one_fm.tasks.erpnext.user.after_insert",
 	},
-	"Operations Post": {
-		"before_save": "one_fm.events.operations_post.before_save",
-	},
 	"Email Queue": {
 		"after_insert":"one_fm.events.email_queue.after_insert",
 	},
